Refactored the functions in settings
