+++ src/module/settings.ts
@@ -1,20 +1,78 @@
-import {localize, MODULE_NAME} from "./util";
+'use strict';
 
-const ORIGIN_FOLDER = 'data';
-const UPLOAD_FOLDER_PATH = 'uploaded-chat-images';
+import {localize, log, MODULE_NAME} from "./utils";
 
-const registerSetting = (setting: any): void => game?.settings?.register(MODULE_NAME, setting.key, setting.options);
+// the path of the foundry data folder
+const ORIGIN_FOLDER = 'data',
+    // the path of the folder where all the uploaded images are saved
+    UPLOAD_FOLDER_PATH = `uploaded-${MODULE_NAME}`;
 
-const getFolder = (source: string, folder: string): Promise<void> => FilePicker.browse(source, folder);
-const createFolder = (source: string, folder: string, options: {}) => FilePicker.createDirectory(source, folder, options);
-const createUploadFolder = (location: any): Promise<any> => location.target === '.' && createFolder(ORIGIN_FOLDER, UPLOAD_FOLDER_PATH, {});
+/**
+ * Browse files for a certain directory location
+ *
+ * @param {string} source - The source location in which to browse. See FilePicker#sources for details
+ * @param {string} target - The target within the source location
+ *
+ * @return {Promise<Object>} A Promise which resolves to the directories and files contained in the location
+ */
+function getFolder(source, target) {
+    return FilePicker.browse(source, target)
+}
 
-// Get the value for a registered setting
-const getSetting = (key: string): any => game?.settings?.get(MODULE_NAME, key);
+/**
+ * Create a subdirectory within a given source. The requested subdirectory path must not already exist.
+ *
+ * @param {string} source - The source location in which to browse
+ * @param {string} target - The target within the source location
+ * @param {Object} [options={}] - Optional arguments which modify the request
+ *
+ * @return {Promise<Object>}
+ */
+function createFolder(source, target, options = {}) {
+    return FilePicker.createDirectory(source, target, options);
+}
 
-// Register all the settings in the array SETTINGS
-const registerSettings = (): void => {
-    const SETTINGS = [
+/**
+ * Check if the uploaded-chat-images folder exists, if not create it. This check runs every time the module is
+ * initialized, just in case the user deletes the folder (or black magic happens and the folder disappears)
+ *
+ * @return {Promise<Object>}
+ */
+function createUploadFolderIfMissing() {
+    return getFolder(ORIGIN_FOLDER, UPLOAD_FOLDER_PATH)
+        .then(location => location.target === '.' && createFolder(ORIGIN_FOLDER, UPLOAD_FOLDER_PATH))
+        .catch(err => log(err));
+}
+
+/**
+ * The function returns the value for a register setting
+ *
+ * @param {string} key - the name of the setting
+ *
+ * @return {*}
+ */
+function getSetting(key) {
+    return game.settings.get(MODULE_NAME, key);
+}
+
+/**
+ * The function calls the foundry register setting function with the module name
+ *
+ * @param {Object} setting
+ * @param {string} setting.key - the name of the setting
+ * @param {Object} setting.options - the properties of the setting
+ */
+function registerSetting(setting) {
+    return game.settings.register(MODULE_NAME, setting.key, setting.options);
+}
+
+/**
+ * Register all the module's settings
+ */
+function registerSettings() {
+    // a list of all the module's settings
+    // initialized here to be able to localize the strings
+    const settings = [
         {
             key: "warningOnDrop",
             options: {
@@ -99,11 +157,7 @@ const registerSettings = (): void => {
         },
     ];
 
-    SETTINGS.forEach(registerSetting);
+    settings.forEach(registerSetting);
 }
-
-// Check if the uploaded-chat-images folder exists, if not create it. This check runs every time the module is
-// initialized, just in case the user deletes the folder (or black magic happens and the folder disappears)
-const createUploadFolderIfMissing = (): Promise<any> => getFolder(ORIGIN_FOLDER, UPLOAD_FOLDER_PATH).then(createUploadFolder);
 
 export {UPLOAD_FOLDER_PATH, getSetting, registerSettings, createUploadFolderIfMissing};
